delegated request input extraction and validation to a separate class
