+++ src/main/java/no/sikt/nva/pubchannels/handler/FetchJournalRequest.java
@@ -1,0 +1,75 @@
+package no.sikt.nva.pubchannels.handler;
+
+import static nva.commons.core.attempt.Try.attempt;
+import java.util.UUID;
+import nva.commons.apigateway.RequestInfo;
+import nva.commons.apigateway.exceptions.BadRequestException;
+import nva.commons.core.StringUtils;
+
+public final class FetchJournalRequest {
+    private static final String IDENTIFIER_PATH_PARAM_NAME = "identifier";
+    private static final String YEAR_PATH_PARAM_NAME = "year";
+    public static final String INVALID_YEAR_MESSAGE = "Invalid path parameter (year). Must be an integer between 1900"
+                                                      + " and 2999.";
+    private static final int MIN_ACCEPTABLE_YEAR = 1900;
+    private static final int MAX_ACCEPTABLE_YEAR = 2999;
+    private static final String INVALID_IDENTIFIER_MESSAGE = "Invalid path parameter (identifier). Must be a UUID "
+                                                             + "version 4.";
+    private static final int UUID_TYPE_4_LENGTH = 36;
+    private static final int YEAR_LENGTH = 4;
+
+
+    private final String identifier;
+    private final String year;
+
+    public FetchJournalRequest(RequestInfo requestInfo) throws BadRequestException {
+        this.identifier = requestInfo.getPathParameter(IDENTIFIER_PATH_PARAM_NAME).trim();
+        this.year = requestInfo.getPathParameter(YEAR_PATH_PARAM_NAME).trim();
+
+        validate();
+    }
+
+    public String getIdentifier() {
+        return identifier;
+    }
+
+    public String getYear() {
+        return year;
+    }
+
+    private void validate() throws BadRequestException {
+        validateIdentifier();
+        validateYear();
+    }
+
+    private void validateYear() throws BadRequestException {
+        if (StringUtils.isEmpty(year)) {
+            throw new BadRequestException(INVALID_YEAR_MESSAGE);
+        }
+
+        if (year.length() != YEAR_LENGTH) {
+            throw new BadRequestException(INVALID_YEAR_MESSAGE);
+        }
+
+        var yearAsInteger = attempt(() -> Integer.parseInt(year))
+                                .orElseThrow(failure -> new BadRequestException(INVALID_YEAR_MESSAGE));
+
+        if (yearAsInteger < MIN_ACCEPTABLE_YEAR || yearAsInteger > MAX_ACCEPTABLE_YEAR) {
+            throw new BadRequestException(INVALID_YEAR_MESSAGE);
+        }
+    }
+
+    private void validateIdentifier() throws BadRequestException {
+        if (StringUtils.isEmpty(identifier)) {
+            throw new BadRequestException(INVALID_IDENTIFIER_MESSAGE);
+        }
+
+        if (identifier.length() != UUID_TYPE_4_LENGTH) {
+            throw new BadRequestException(INVALID_IDENTIFIER_MESSAGE);
+        }
+
+        attempt(() -> UUID.fromString(identifier))
+            .orElseThrow(failure -> new BadRequestException(INVALID_IDENTIFIER_MESSAGE));
+    }
+
+}
